[node] presubstitute
